Revert "feat: extract generic payment logic (#1693)"
This reverts commit 1003b98919ad233d32a3cd083044f21d3295ef34.
