+++ core/meterer/meterer.go
@@ -2,11 +2,14 @@ package meterer
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"math"
+	"math/big"
+	"slices"
 	"time"
 
 	"github.com/Layr-Labs/eigenda/core"
-	"github.com/Layr-Labs/eigenda/core/meterer/payment_logic"
 	"github.com/Layr-Labs/eigensdk-go/logging"
 )
 
@@ -81,7 +84,7 @@ func (m *Meterer) MeterRequest(ctx context.Context, header core.PaymentMetadata,
 		return nil, fmt.Errorf("failed to get payment global params: %w", err)
 	}
 	// Validate against the payment method
-	if !payment_logic.IsOnDemandPayment(&header) {
+	if !IsOnDemandPayment(&header) {
 		reservations, err := m.ChainPaymentState.GetReservedPaymentByAccountAndQuorums(ctx, header.AccountID, quorumNumbers)
 		if err != nil {
 			return nil, fmt.Errorf("failed to get active reservation by account: %w", err)
@@ -115,7 +118,7 @@ func (m *Meterer) serveReservationRequest(
 	receivedAt time.Time,
 ) (map[core.QuorumID]uint64, error) {
 	m.logger.Debug("Recording and validating reservation usage", "header", header, "reservation", reservations)
-	if err := payment_logic.ValidateReservations(reservations, globalParams.QuorumProtocolConfigs, quorumNumbers, header.Timestamp, receivedAt.UnixNano()); err != nil {
+	if err := ValidateReservations(reservations, globalParams.QuorumProtocolConfigs, quorumNumbers, header.Timestamp, receivedAt.UnixNano()); err != nil {
 		return nil, fmt.Errorf("invalid reservation: %w", err)
 	}
 
@@ -143,10 +146,10 @@ func (m *Meterer) incrementBinUsage(
 		if err != nil {
 			return nil, fmt.Errorf("failed to get quorum config for quorum %d: %w", quorumID, err)
 		}
-		charges[quorumID] = payment_logic.SymbolsCharged(numSymbols, protocolConfig.MinNumSymbols)
+		charges[quorumID] = SymbolsCharged(numSymbols, protocolConfig.MinNumSymbols)
 		quorumNumbers = append(quorumNumbers, quorumID)
 		reservationWindows[quorumID] = protocolConfig.ReservationRateLimitWindow
-		requestReservationPeriods[quorumID] = payment_logic.GetReservationPeriodByNanosecond(header.Timestamp, protocolConfig.ReservationRateLimitWindow)
+		requestReservationPeriods[quorumID] = GetReservationPeriodByNanosecond(header.Timestamp, protocolConfig.ReservationRateLimitWindow)
 	}
 	// Batch increment all quorums for the current quorums' reservation period
 	// For each quorum, increment by its specific symbolsCharged value
@@ -160,7 +163,7 @@ func (m *Meterer) incrementBinUsage(
 	for quorumID, reservation := range reservations {
 		reservationWindow := reservationWindows[quorumID]
 		requestReservationPeriod := requestReservationPeriods[quorumID]
-		usageLimit := payment_logic.GetBinLimit(reservation.SymbolsPerSecond, reservationWindow)
+		usageLimit := GetReservationBinLimit(reservation, reservationWindow)
 		newUsage, ok := usage[quorumID]
 		if !ok {
 			return nil, fmt.Errorf("failed to get updated usage for quorum %d", quorumID)
@@ -172,8 +175,8 @@ func (m *Meterer) incrementBinUsage(
 			// Bin was already filled before this increment
 			return nil, fmt.Errorf("bin has already been filled for quorum %d", quorumID)
 		}
-		overflowPeriod := payment_logic.GetOverflowPeriod(requestReservationPeriod, reservationWindow)
-		if charges[quorumID] <= usageLimit && overflowPeriod <= payment_logic.GetReservationPeriod(int64(reservation.EndTimestamp), reservationWindow) {
+		overflowPeriod := GetOverflowPeriod(requestReservationPeriod, reservationWindow)
+		if charges[quorumID] <= usageLimit && overflowPeriod <= GetReservationPeriod(int64(reservation.EndTimestamp), reservationWindow) {
 			// Needs to go to overflow bin
 			overflowAmounts[quorumID] = newUsage - usageLimit
 			overflowPeriods[quorumID] = overflowPeriod
@@ -211,7 +214,7 @@ func (m *Meterer) incrementBinUsage(
 func (m *Meterer) serveOnDemandRequest(ctx context.Context, globalParams *PaymentVaultParams, header core.PaymentMetadata, onDemandPayment *core.OnDemandPayment, numSymbols uint64, headerQuorums []uint8, receivedAt time.Time) (map[core.QuorumID]uint64, error) {
 	m.logger.Debug("Recording and validating on-demand usage", "header", header, "onDemandPayment", onDemandPayment)
 
-	if err := payment_logic.ValidateQuorum(headerQuorums, globalParams.OnDemandQuorumNumbers); err != nil {
+	if err := ValidateQuorum(headerQuorums, globalParams.OnDemandQuorumNumbers); err != nil {
 		return nil, fmt.Errorf("invalid quorum for On-Demand Request: %w", err)
 	}
 
@@ -225,8 +228,8 @@ func (m *Meterer) serveOnDemandRequest(ctx context.Context, globalParams *Paymen
 		return nil, fmt.Errorf("failed to get payment config for on-demand quorum: %w", err)
 	}
 
-	symbolsCharged := payment_logic.SymbolsCharged(numSymbols, protocolConfig.MinNumSymbols)
-	paymentCharged := payment_logic.PaymentCharged(symbolsCharged, paymentConfig.OnDemandPricePerSymbol)
+	symbolsCharged := SymbolsCharged(numSymbols, protocolConfig.MinNumSymbols)
+	paymentCharged := PaymentCharged(symbolsCharged, paymentConfig.OnDemandPricePerSymbol)
 	oldPayment, err := m.MeteringStore.AddOnDemandPayment(ctx, header, paymentCharged)
 	if err != nil {
 		return nil, fmt.Errorf("failed to update cumulative payment: %w", err)
@@ -259,14 +262,157 @@ func (m *Meterer) incrementGlobalBinUsage(ctx context.Context, params *PaymentVa
 		return fmt.Errorf("failed to get quorum configs for on-demand quorum: %w", err)
 	}
 
-	globalPeriod := payment_logic.GetReservationPeriod(receivedAt.Unix(), protocolConfig.OnDemandRateLimitWindow)
+	globalPeriod := GetReservationPeriod(receivedAt.Unix(), protocolConfig.OnDemandRateLimitWindow)
 
 	newUsage, err := m.MeteringStore.UpdateGlobalBin(ctx, globalPeriod, symbolsCharged)
 	if err != nil {
 		return fmt.Errorf("failed to increment global bin usage: %w", err)
 	}
-	if newUsage > payment_logic.GetBinLimit(paymentConfig.OnDemandSymbolsPerSecond, protocolConfig.OnDemandRateLimitWindow) {
+	if newUsage > GetBinLimit(paymentConfig.OnDemandSymbolsPerSecond, protocolConfig.OnDemandRateLimitWindow) {
 		return fmt.Errorf("global bin usage overflows")
 	}
 	return nil
+}
+
+// GetReservationBinLimit returns the bin limit for a given reservation
+// Note: This is called per-quorum since reservation is for a single quorum.
+func GetReservationBinLimit(reservation *core.ReservedPayment, reservationWindow uint64) uint64 {
+	return GetBinLimit(reservation.SymbolsPerSecond, reservationWindow)
+}
+
+// GetBinLimit returns the bin limit given the bin interval and the symbols per second
+func GetBinLimit(symbolsPerSecond uint64, binInterval uint64) uint64 {
+	return symbolsPerSecond * binInterval
+}
+
+// GetReservationPeriodByNanosecondTimestamp returns the current reservation period by finding the nearest lower multiple of the bin interval;
+// bin interval used by the disperser is publicly recorded on-chain at the payment vault contract
+func GetReservationPeriodByNanosecond(nanosecondTimestamp int64, binInterval uint64) uint64 {
+	if nanosecondTimestamp < 0 {
+		return 0
+	}
+	return GetReservationPeriod(int64((time.Duration(nanosecondTimestamp) * time.Nanosecond).Seconds()), binInterval)
+}
+
+// GetReservationPeriod returns the current reservation period by finding the nearest lower multiple of the bin interval;
+// bin interval used by the disperser is publicly recorded on-chain at the payment vault contract
+func GetReservationPeriod(timestamp int64, binInterval uint64) uint64 {
+	if binInterval == 0 {
+		return 0
+	}
+	return uint64(timestamp) / binInterval * binInterval
+}
+
+// GetOverflowPeriod returns the overflow period by adding the overflow offset to the current reservation period
+// the offset is 2*reservationWindow, skipping the immediate next period for the period that will be used for overflow from the current period
+func GetOverflowPeriod(reservationPeriod uint64, reservationWindow uint64) uint64 {
+	return reservationPeriod + reservationWindow*2
+}
+
+// PaymentCharged returns the chargeable price for a given number of symbols
+func PaymentCharged(numSymbols, pricePerSymbol uint64) *big.Int {
+	// directly convert to uint64 to avoid overflow
+	numSymbolsInt := new(big.Int).SetUint64(numSymbols)
+	pricePerSymbolInt := new(big.Int).SetUint64(pricePerSymbol)
+	return new(big.Int).Mul(numSymbolsInt, pricePerSymbolInt)
+}
+
+// SymbolsCharged returns the number of symbols charged for a given data length
+// being at least MinNumSymbols or the nearest rounded-up multiple of MinNumSymbols.
+func SymbolsCharged(numSymbols uint64, minSymbols uint64) uint64 {
+	if numSymbols <= minSymbols {
+		return minSymbols
+	}
+	if minSymbols == 0 {
+		return numSymbols
+	}
+	// Round up to the nearest multiple of MinNumSymbols
+	roundedUp := core.RoundUpDivide(numSymbols, minSymbols) * minSymbols
+	// Check for overflow; this case should never happen
+	if roundedUp < numSymbols {
+		return math.MaxUint64
+	}
+	return roundedUp
+}
+
+// ValidateQuorum ensures that the quorums listed in the blobHeader are present within allowedQuorums
+// Note: A reservation that does not utilize all of the allowed quorums will be accepted. However, it
+// will still charge against all of the allowed quorums. A on-demand requests require and only allow
+// the ETH and EIGEN quorums.
+func ValidateQuorum(headerQuorums []uint8, allowedQuorums []uint8) error {
+	if len(headerQuorums) == 0 {
+		return fmt.Errorf("no quorum numbers provided in the request")
+	}
+
+	// check that all the quorum ids are in ReservedPayment's
+	for _, q := range headerQuorums {
+		if !slices.Contains(allowedQuorums, q) {
+			// fail the entire request if there's a quorum number mismatch
+			return fmt.Errorf("quorum number mismatch: %d", q)
+		}
+	}
+	return nil
+}
+
+// ValidateReservations ensures that the quorums listed in the blobHeader are present within allowedQuorums.
+//
+// Notes:
+//   - Reservations that don't use all allowed quorums are still accepted
+//   - Charges apply to ALL allowed quorums, even if not all are used
+//   - On-demand requests have special requirements: they must use ETH and EIGEN quorums only
+func ValidateReservations(
+	reservations map[core.QuorumID]*core.ReservedPayment,
+	quorumConfigs map[core.QuorumID]*core.PaymentQuorumProtocolConfig,
+	quorumNumbers []uint8,
+	paymentHeaderTimestampNs int64,
+	receivedTimestampNs int64,
+) error {
+	reservationQuorums := make([]uint8, 0, len(reservations))
+	reservationWindows := make(map[core.QuorumID]uint64, len(reservations))
+	requestReservationPeriods := make(map[core.QuorumID]uint64, len(reservations))
+
+	// Gather quorums the user had an reservations on and relevant quorum configurations
+	for quorumID := range reservations {
+		reservationQuorums = append(reservationQuorums, uint8(quorumID))
+		_, ok := quorumConfigs[quorumID]
+		if !ok {
+			return fmt.Errorf("quorum config not found for quorum %d", quorumID)
+		}
+		reservationWindows[quorumID] = quorumConfigs[quorumID].ReservationRateLimitWindow
+		requestReservationPeriods[quorumID] = GetReservationPeriodByNanosecond(paymentHeaderTimestampNs, quorumConfigs[quorumID].ReservationRateLimitWindow)
+	}
+	if err := ValidateQuorum(quorumNumbers, reservationQuorums); err != nil {
+		return err
+	}
+	// Validate the used reservations are active and is of valid periods
+	for _, quorumID := range quorumNumbers {
+		reservation := reservations[core.QuorumID(quorumID)]
+		if !reservation.IsActiveByNanosecond(paymentHeaderTimestampNs) {
+			return errors.New("reservation not active")
+		}
+		if !ValidateReservationPeriod(reservation, requestReservationPeriods[quorumID], reservationWindows[quorumID], receivedTimestampNs) {
+			return fmt.Errorf("invalid reservation period for reservation on quorum %d", quorumID)
+		}
+	}
+
+	return nil
+}
+
+// ValidateReservationPeriod checks if the provided reservation period is valid
+// Note: This is called per-quorum since reservation is for a single quorum.
+func ValidateReservationPeriod(reservation *core.ReservedPayment, requestReservationPeriod uint64, reservationWindow uint64, receivedTimestampNs int64) bool {
+	currentReservationPeriod := GetReservationPeriodByNanosecond(receivedTimestampNs, reservationWindow)
+	// Valid reservation periods are either the current bin or the previous bin
+	isCurrentOrPreviousPeriod := requestReservationPeriod == currentReservationPeriod || requestReservationPeriod == (currentReservationPeriod-reservationWindow)
+	startPeriod := GetReservationPeriod(int64(reservation.StartTimestamp), reservationWindow)
+	endPeriod := GetReservationPeriod(int64(reservation.EndTimestamp), reservationWindow)
+	isWithinReservationWindow := startPeriod <= requestReservationPeriod && requestReservationPeriod < endPeriod
+	if !isCurrentOrPreviousPeriod || !isWithinReservationWindow {
+		return false
+	}
+	return true
+}
+
+func IsOnDemandPayment(paymentMetadata *core.PaymentMetadata) bool {
+	return paymentMetadata.CumulativePayment.Cmp(big.NewInt(0)) > 0
 }
