+++ core/meterer/period_record_test.go
@@ -6,7 +6,6 @@ import (
 	disperser_rpc "github.com/Layr-Labs/eigenda/api/grpc/disperser/v2"
 	"github.com/Layr-Labs/eigenda/core"
 	"github.com/Layr-Labs/eigenda/core/meterer"
-	"github.com/Layr-Labs/eigenda/core/meterer/payment_logic"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -373,8 +372,8 @@ func TestQuorumPeriodRecords_UpdateUsage(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			// Calculate expected periods for setup
-			currentPeriod := payment_logic.GetReservationPeriodByNanosecond(tt.timestamp, tt.protocolConfig.ReservationRateLimitWindow)
-			overflowPeriod := payment_logic.GetOverflowPeriod(currentPeriod, tt.protocolConfig.ReservationRateLimitWindow)
+			currentPeriod := meterer.GetReservationPeriodByNanosecond(tt.timestamp, tt.protocolConfig.ReservationRateLimitWindow)
+			overflowPeriod := meterer.GetOverflowPeriod(currentPeriod, tt.protocolConfig.ReservationRateLimitWindow)
 
 			// Setup initial records if needed
 			if tt.setupCurrentRecord {
